Spell bar now saves/restores from db. (Fixed for EF)

+++ Source/ACE.Server/Network/Handlers/CharacterHandler.cs
@@ -432,15 +432,15 @@ namespace ACE.Server.Network.Handlers
                         // Olthoi Spitter is a special case
                         if (characterCreateInfo.Heritage == (int)HeritageGroup.OlthoiAcid)
                         {
-                            player.AddKnownSpell((int)spell.SpellId);
+                            player.AddKnownSpell(spell.SpellId);
                             // Continue to next spell as Olthoi spells do not have the SpecializedOnly field
                             continue;
                         }
 
                         if (charSkill.Status == SkillStatus.Trained && spell.SpecializedOnly == false)
-                            player.AddKnownSpell((int)spell.SpellId);
+                            player.AddKnownSpell(spell.SpellId);
                         else if (charSkill.Status == SkillStatus.Specialized)
-                            player.AddKnownSpell((int)spell.SpellId);
+                            player.AddKnownSpell(spell.SpellId);
                     }
                 }
             }
